+++ smash-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for possibly-can
+## 0.1.1
 
-## 0.1.0.0 -- YYYY-mm-dd
+* Add `NFData`, `Binary` instances
+* CPP to extend to 8.2.2 without warnings
+
+## 0.1.0.0
 
 * First version. Released on an unsuspecting world.